backend/requirements.txt
asgiref==3.2.3
Django==3.0.3
django-cors-headers==3.2.1
django-debug-toolbar==2.1
django-extensions==2.2.8
django-shell-plus==1.1.7
djangorestframework==3.10.3
python-decouple==3.3
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
